+++ lib/src/context_extensions.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart' show Clipboard, ClipboardData;
 
 /// Extension methods on BuildContext
 extension ContextExtension on BuildContext {
@@ -139,93 +138,4 @@ extension ContextExtension on BuildContext {
   ///and push the [path] using Navigator 1.0
   ///
   void popAndPushNamed(String path) => Navigator.of(this).popAndPushNamed(path);
-}
-
-extension StringExtension on String {
-  ///Extension method on String to capitalise first character of the string.
-  ///
-  String capitaliseFirstChar() =>
-      isEmpty ? this : this[0].toUpperCase() + substring(1);
-
-  ///Extension method on String to capitalise first character
-  ///of each word of the string.
-  ///
-  String capitaliseEachWordFirstChar() => isEmpty
-      ? this
-      : replaceAll(RegExp(' +'), ' ')
-          .split(' ')
-          .map((str) => str.capitaliseFirstChar())
-          .join(' ');
-
-  /// Extension method to copy a string to clipboard.
-  /// It takes no parameters.
-  /// This is how you can use it:
-  /// ```dart
-  /// String text = "The beautiful thing with Dart is that you  can create your own extensions.";
-  ///  TextButton(
-  ///     child: const Text('Copy to clipboard'),
-  ///     onPressed: () => text.copyToClipboard(),
-  ///    );
-  /// ```
-  Future<void> copyToClipboard() =>
-      Clipboard.setData(ClipboardData(text: this));
-
-  /// Extension method to check if a string is a valid email.
-  /// You can pass a custom [emailPattern] as a parameter.
-  /// This is how you can use it:
-  /// ```dart
-  /// String email = 'email@example.com';
-  ///   final bool isValidEmail = email.isEmailValid();
-  ///
-  /// or pass a custom pattern as a parameter like this:
-  ///   final bool isValidEmail = email.isEmailValid(
-  ///     emailPattern: r'^[a-zA-Z0-9.]+@[a-zA-Z0-9]+\.[a-zA-Z]+',
-  ///   );
-  /// ```
-  bool isValidEmail({Pattern? emailPattern}) {
-    final Pattern pattern = emailPattern ??
-        r'^(([^<>()[\]\\.,;:\s@\"]+(\.[^<>()[\]\\.,;:\s@\"]+)*)|(\".+\"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$';
-
-    final RegExp regex = RegExp(pattern as String);
-    return isNotEmpty && regex.hasMatch(trim());
-  }
-
-  /// Extension method to get the initials of a string.
-  /// It takes no parameters.
-  /// This is how you can use it:
-  /// ```dart
-  /// String text = "Example Name";
-  /// TextButton(
-  ///    child: const Text('Get initials'),
-  ///    onPressed: () => text.getInitials(),
-  ///    );
-  ///    ```
-  ///    It returns "En" in this case.
-  ///    If the string is empty, it returns null.
-  ///    If the string has only one word, it returns the first character of the word.
-  ///    If the string has more than one word, it returns the first character of each word.
-  ///    ```dart
-  String? get getInitials {
-    final trimmed = trim();
-    if (trimmed.isEmpty) {
-      return null;
-    }
-    final parts = trimmed.split(' ');
-    if (parts.length == 1) {
-      return parts.first.characters.first;
-    } else {
-      final firstLetters = parts.map((part) => part.characters.first).join();
-      return firstLetters;
-    }
-  }
-}
-
-extension NumberExtension on num {
-  /// Creates an empty SizedBox with height as the given number.
-  ///
-  SizedBox get hSizedBox => SizedBox(height: toDouble());
-
-  /// Creates an empty SizedBox with width as the given number.
-  ///
-  SizedBox get wSizedBox => SizedBox(width: toDouble());
 }
