+++ lib/super_extensions.dart
@@ -1,1 +1,3 @@
-export 'src/super_extensions_base.dart';
+export 'src/context_extensions.dart';
+export 'src/number_extensions.dart';
+export 'src/string_extensions.dart';
